+++ Arbiter.Net/Client/ClientMetadataRequestType.cs
@@ -1,7 +1,0 @@
-﻿namespace Arbiter.Net.Client;
-
-public enum ClientMetadataRequestType : byte
-{
-    GetMetadata = 0,
-    Listing = 1
-}