+++ Arbiter.Net/Client/Messages/ClientRequestMetadataMessage.cs
@@ -1,19 +1,20 @@
 ﻿using Arbiter.Net.Annotations;
 using Arbiter.Net.Serialization;
+using Arbiter.Net.Types;
 
 namespace Arbiter.Net.Client.Messages;
 
 [NetworkCommand(ClientCommand.RequestMetadata)]
 public class ClientRequestMetadataMessage : ClientMessage
 {
-    public ClientMetadataRequestType RequestType { get; set; }
+    public MetadataRequestType RequestType { get; set; }
     public string? Name { get; set; }
 
     public override void Deserialize(INetworkPacketReader reader)
     {
         base.Deserialize(reader);
         
-        RequestType = (ClientMetadataRequestType)reader.ReadByte();
+        RequestType = (MetadataRequestType)reader.ReadByte();
         Name = reader.ReadString8();
     }