+++ Arbiter.Net/Types/MetadataRequestType.cs
@@ -1,0 +1,7 @@
+﻿namespace Arbiter.Net.Types;
+
+public enum MetadataRequestType : byte
+{
+    GetMetadata = 0,
+    Listing = 1
+}