+++ CHANGELOG.md
@@ -5,9 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+
+### Changed
+
+- Client name filter now matches exactly (unless `*` or `?` are used)
+
 ## [0.9.3] - 2025-09-18
 
-## Added
+### Added
 
 - `Copy Selected` context menu for trace packets
 - `Save Selected` context menu for trace packets
@@ -15,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Double-click on client in list to bring game window to foreground (Windows only)
 - Set game client window title to include client name on login (cleared on logout)
 
-## Changed
+### Changed
 
 - Allow multiple selection of trace packets (only first item is inspected)
 - Fixed `Shift` modifier getting stuck when loading traces in append mode